feat: add PARTS constant and part range validation
- Add const PARTS: u8 to Solver trait for compile-time part count
- Add SolverExt trait with solve_part_checked_range()
- Change year/day/part types from u32/u32/usize to u16/u8/u8
- Add property-based tests for part bounds validation

+++ aoc-solver/src/error.rs
@@ -21,10 +21,10 @@ pub enum ParseError {
 pub enum SolveError {
     /// The requested part number is not implemented
     #[error("Part {0} is not implemented")]
-    PartNotImplemented(usize),
+    PartNotImplemented(u8),
     /// The requested part number is out of range (exceeds max_parts)
     #[error("Part {0} is out of range")]
-    PartOutOfRange(usize),
+    PartOutOfRange(u8),
     /// An error occurred while solving the part
     #[error("Solve failed: {0}")]
     SolveFailed(#[source] Box<dyn std::error::Error + Send + Sync>),
@@ -35,7 +35,7 @@ pub enum SolveError {
 pub enum SolverError {
     /// Solver not found for the given year and day
     #[error("Solver not found for year {0} day {1}")]
-    NotFound(u32, u32),
+    NotFound(u16, u8),
     /// Error occurred during parsing
     #[error("Parse error: {0}")]
     ParseError(#[from] ParseError),
@@ -49,5 +49,5 @@ pub enum SolverError {
 pub enum RegistrationError {
     /// Attempted to register a solver for a year-day combination that already exists
     #[error("Duplicate solver registration for year {0} day {1}")]
-    DuplicateSolver(u32, u32),
+    DuplicateSolver(u16, u8),
 }
